Calibrated both the Pressure and temperature readings. In addition, both the humidity and temperature readings were also calibrated.

+++ aws-iot-ethernet/c14227/project/git/aws-dev/software/aws-iot-ethernet/firmware/src/app.h
@@ -375,7 +375,7 @@ typedef struct
 
 	int I2C1_val1,I2C1_val2;
 	long int pressure_value;										// Declaration of pressure variable in mbar
-	int temperature_value;										// Declaration of temperature variable
+	float temperature_value;										// Declaration of temperature variable
 	unsigned char chip_id_address;
 	float Temperature;
 	float Humidity;
